Fix long name #75

+++ FtpServer.cpp
@@ -1918,7 +1918,7 @@ void FtpServer::abortTransfer()
 
 int32_t FtpServer::readChar()
 {
-  int8_t rc = -1;
+  int32_t rc = -1;
 
   if( client.available())
   {
